Check for existing post before deleting

+++ server/controllers/post/deleteById.js
@@ -1,6 +1,9 @@
 // database
 const database = require("../../database");
 
+// services
+const getPostById = require("../../services/posts/getPostById");
+
 // utils
 const logger = require("../../utils/logger");
 const error = require("../../errorResponse.json");
@@ -17,6 +20,16 @@ exports.deleteById = async (req, res) => {
 	}
 
 	try {
+		const getPost = await getPostById(postId);
+
+		if (!getPost) {
+			res.status(404).send({
+				message: error.api.posts.postNotFound,
+				code: "POST_NOT_FOUND"
+			});
+			return;
+		}
+
 		await database
 			.delete()
 			.from("posts")
